Fix: let carousel item be optimized
